+++ CHANGELOG.md
@@ -196,6 +196,7 @@
 * Added some handling for fatal errors during startup
 * Added editor script that adds build info to build
 * ActorAttackComponent now derives from abstract base class
+* ActorAttackComponentBase derivatives can optionally handle target selection and chase destination
 * Added support for glue-to-bone to ranged weapon viewmodel [Experimental]
 * Added contact-hit hack to melee weapons [Experimental]
 * Implemented HUD scaling [Experimental]